backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.0
azure-cosmos==4.6.0
azure-identity==1.17.1
python-dotenv==1.0.1
openai==1.12.0
httpx==0.24.1
httpcore==0.17.3
gunicorn==21.2.0